feat: add hoo useWindowWidth
