+++ src/containers/Header/Header.styles.ts
@@ -1,35 +1,22 @@
 import styled from 'styled-components';
 
 export const Header = styled.header`
-  max-width: 100%;
+  width: 100%;
   height: 78px;
-
+  position: fixed;
   display: flex;
-  justify-content: space-between;
   align-items: center;
   padding: 0 20px;
-  border-color: transparent;
-  gap: 32px;
+  box-shadow: 0 0.4rem 0.8rem 0 rgba(99, 99, 99, 0.2);
+
+  @media (max-width: 768px) {
+    justify-content: space-between;
+  }
 `;
 
 export const Logo = styled.div`
+  min-width: 33%;
   color: ${({ theme }) => theme.colors.neutral.medium};
   font-size: ${({ theme }) => theme.fonts.size.md};
   font-weight: bold;
-`;
-
-export const Menu = styled.nav`
-  display: flex;
-  gap: 32px;
-  align-items: center;
-  font-weight: bolder;
-
-  :hover {
-    color: ${({ theme }) => theme.colors.neutral.medium};
-  }
-
-  @media (max-width: 540px) {
-    align-items: flex-start;
-    flex-direction: column;
-  }
 `;
