+++ src/containers/Header/Header.tsx
@@ -1,54 +1,23 @@
 import { ModalContent } from '../../componentes/Modal';
+import { useWindowWidth } from '../../hooks/useWindowWidth';
 import * as S from './Header.styles';
-import { useState, useEffect } from 'react';
+
+import { NavBar } from './components/NavBar';
 
 export const Header = () => {
-  const [windowWidth, setWindowWidth] = useState(window.innerWidth);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setWindowWidth(window.innerWidth);
-    };
-
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }, []);
+  const windowWidth = useWindowWidth();
 
   return (
     <S.Header>
       <S.Logo>Viní Viseman</S.Logo>
-      {windowWidth > 540 ? (
-        <S.Menu>
-          <a href="/">Home</a>
 
-          <a href="/">Sobre</a>
-
-          <a href="/">Blog</a>
-
-          <a href="/">Soluções</a>
-
-          <a href="/">Contato</a>
-        </S.Menu>
+      {windowWidth > 768 ? (
+        <NavBar />
       ) : (
         <ModalContent>
-          <S.Menu>
-            <a href="/">Home</a>
-
-            <a href="/">Sobre</a>
-
-            <a href="/">Blog</a>
-
-            <a href="/">Soluções</a>
-
-            <a href="/">Contato</a>
-          </S.Menu>
+          <NavBar />
         </ModalContent>
       )}
-
-      <button>Teste</button>
     </S.Header>
   );
 };
