+++ src/containers/Header/components/NavBar.tsx
@@ -1,0 +1,20 @@
+import * as S from './NavBar.styles';
+
+export const NavBar = () => {
+  return (
+    <S.NavBar>
+      <S.NavBarContent>
+        <a href="/">Home</a>
+
+        <a href="/">Sobre</a>
+
+        <a href="/">Blog</a>
+
+        <a href="/">Soluções</a>
+
+        <a href="/">Contato</a>
+      </S.NavBarContent>
+      <button>Teste</button>
+    </S.NavBar>
+  );
+};
