+++ src/containers/Header/index.ts
@@ -1,1 +1,1 @@
-export * from './Header'
+export * from './Header';
